Attempt to simplify WordManager.
* Extracted the loading of word list to a separate class
* Limited the number of different ways of storing words, since they
  can be derived from each others
* Moved to immutable data structure (one step closer to thread-safety)

This refactoring shows that there are still issues with thread-safety
and with the management of `null`. A number of methods, parameters and
fields have been tagged with @Nullable and @Nonnull to clarify the
current situation.

The constructor has some problematic exception handling, which might
lead to a half-initialized instance. A better understanding of the
plugin lifecycle is needed to propose a coherent solution.